server/requirements.txt
fastapi
uvicorn
python-dotenv
torch
diffusers
transformers
pydantic
pillow
psutil
huggingface-hub
safetensors
imageio
numpy